move add task modal to separate folder
